my?lnik jest.

+++ src/Engine/Models/Resources/Font/TextHelper.cpp
@@ -193,7 +193,7 @@ float                    TextHelper::BuildVACForText     ( VertexAttributesChann
 
         if( auto glyphCoord = textAtlas->GetGlyphCoords( wch ) )
         {
-            glm::vec3 baring = glm::vec3( 0.f, (glyphCoord->glyphHeight - glyphCoord->bearingY) / (float)viewHeight, 0.f );
+            glm::vec3 baring = glm::vec3( 0.f, ( (float)glyphCoord->glyphHeight - (float)glyphCoord->bearingY ) / (float)viewHeight, 0.f );
 
             glm::vec3 quadBottomLeft;
             glm::vec3 quadBottomRight;
